refactor xrefs: https://github.com/metanorma/isodoc/issues/606
