+++ lib/isodoc/bipm/base_convert.rb
@@ -27,22 +27,6 @@ module IsoDoc
         end
       end
 
-      def render_identifier(ident)
-        ret = super
-        ret[:sdo] = ret[:sdo]&.sub(/^(BIPM)([  ])(PV|CR)([  ])(\d.*)$/,
-                                   "\\1\\2<strong>\\3</strong>,\\4\\5")
-        ret
-      end
-
-      def implicit_reference(bib)
-        b = bib.at(ns("./docidentifier[@primary = 'true'][@type = 'BIPM']"))
-        return true if @doctype == "brochure" && /^(CGPM|CIPM|CCDS|CCTF)[  ]
-        (Resolution|Recommendation|Declaration|Decision|Recommendation|Meeting)/x
-          .match?(b&.text)
-
-        super
-      end
-
       def term_cleanup(docxml)
         @jcgm ? docxml : super
       end
