+++ lib/isodoc/bipm/init.rb
@@ -11,7 +11,7 @@ module IsoDoc
       end
 
       def xref_init(lang, script, _klass, labels, options)
-        html = HtmlConvert.new(language: lang, script: script)
+        html = PresentationXMLConvert.new(language: lang, script: script)
         @xrefs = Xref.new(lang, script, html, labels, options)
       end
 
@@ -23,11 +23,6 @@ module IsoDoc
       def bibrenderer(options = {})
         ::Relaton::Render::Bipm::General.new(options.merge(language: @lang,
                                                            i18nhash: @i18n.get))
-      end
-
-      def omit_docid_prefix(prefix)
-        %w(BIPM BIPM-long).include? prefix and return true
-        super
       end
 
       def amd?(_text)
