+++ lib/isodoc/bipm/presentation_xml_convert.rb
@@ -6,6 +6,7 @@ require_relative "index"
 require_relative "doccontrol"
 require_relative "../../relaton/render/general"
 require_relative "presentation_blocks"
+require_relative "presentation_biblio"
 
 module IsoDoc
   module Bipm
@@ -102,33 +103,6 @@ module IsoDoc
         super
       end
 
-      def bibdata_i18n(bibdata)
-        super
-        bibdata_dates(bibdata)
-        bibdata_titles(bibdata)
-      end
-
-      def bibdata_dates(bibdata)
-        pubdate = bibdata.at(ns("./date[not(@format)][@type = 'published']"))
-        pubdate or return
-        meta = metadata_init(@lang, @script, @locale, @i18n)
-        pubdate.next = pubdate.dup
-        pubdate.next["format"] = "ddMMMyyyy"
-        pubdate.next.children = meta.monthyr(pubdate.text)
-      end
-
-      def bibdata_titles(bibdata)
-        app = bibdata.at(ns("//bibdata/ext/" \
-                            "structuredidentifier/part")) or return
-        bibdata.xpath(ns("//bibdata/title[@type = 'title-part']")).each do |t|
-          t.previous = t.dup
-          t["type"] = "title-part-with-numbering"
-          part = t["language"] == "en" ? "Part" : "Partie" # not looking up in YAML
-          t.children = l10n("#{part} #{app.text}: #{to_xml(t.children)}",
-                            t["language"])
-        end
-      end
-
       def eref(docxml)
         super
         jcgm_eref(docxml, "//eref")
@@ -210,25 +184,6 @@ module IsoDoc
         # s = termsource_status(elem["status"]) and origin.next = l10n(", #{s}")
         # end
         termsource_add_modification_text(elem.at(ns("./modification")))
-      end
-
-      def norm_ref_entry_code(_ordinal, identifiers, _ids, _standard, datefn,
-_bib)
-        ret = identifiers[0] || identifiers[1]
-        ret += " #{identifiers[1]}" if identifiers[0] && identifiers[1]
-        "#{ret}#{datefn} "
-      end
-
-      def biblio_ref_entry_code(ordinal, ids, _id, _standard, datefn, _bib)
-        # standard and id = nil
-        ret = ids[:ordinal] || ids[:metanorma] || "[#{ordinal}]"
-        if ids[:sdo]
-          ret = prefix_bracketed_ref(ret)
-          ret += "#{ids[:sdo]}#{datefn} "
-        else
-          ret = prefix_bracketed_ref("#{ret}#{datefn}")
-        end
-        ret
       end
 
       include Init
